refine Segment definition; state definition of d-separation
We definitely want to be able to match on segment topology, but it's
also useful to be able to index into the nodes in a segment without
matching on the topology; thus, the topology and node membership tuple
should be separate fields in a struct; the tuple should not be enum
data.

The present author is pleased with the relatively clean appearance of
the d_separated method once the requisite concept definitions are in
place.