draft definition of networks although critical structural problems lurk
I may have misunderstood the limitations of Cameron's
graph-represenation-with-Rc-and-RefCell example; even though Bayes nets
are conceptually DAGs, I still want backpointers to make it easy to
traverse a path against the arrows (as is implied by the definition of
d-separation). But that makes our network datastructure outright cyclic,
which is apparently worse than merely containing non-cylic shared
references. So the code checked-in in this commit is fine, but when
taking the obvious next step of defining

    pub fn link(&self, parent: &Node, child: &Node) {
        if !self.0.contains(&parent) || !self.0.contains(&child) {
            panic!("cannot link nodes absent from network");
        }

        parent.borrow_mut().children.push(child.clone());
        child.borrow_mut().parents.push(parent.clone());
    }

—then we blow the stack.

More research is needed.