schema rewrite: Weak parents; RefCell around fields rather than entirety
We got rid of the Node typealias, but with the RefCells tightly scoped
to the particular fields I expect to mutate rather than the entire
Variable, it doesn't feel too boilerplatey.

We want to call the make-a-new-Variable method something other than
`::new`, because it returns an Rc<Variable> rather than a Variable. Thus,
`::create`.

It didn't really make sense to convert Segments from Paths rather than
just a length-three slice. To avoid type mismatches between Weak and Rc,
Segment::new always accesses the children vector of a Varable.